removed some dead code added an interface for StripedExecutor
